+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import './App.css'
 
-import Dashboard from './Dashboard/DashboardContainer'
+import Dashboard from './Dashboard'
 
 function App () {
   return (
