+++ src/Dashboard/components/ContactList/ContactItem/index.js
@@ -1,0 +1,2 @@
+import ContactItem from './ContactItem'
+export default ContactItem
