+++ src/Dashboard/components/ContactList/InitialLetter/index.js
@@ -1,0 +1,2 @@
+import InitialLetter from './InitialLetter'
+export default InitialLetter
