+++ src/Dashboard/components/ContactList/index.js
@@ -1,0 +1,2 @@
+import ContactList from './ContactList'
+export default ContactList
