+++ src/Dashboard/components/CreateContactForm/index.js
@@ -1,0 +1,2 @@
+import CreateContactForm from './CreateContactForm'
+export default CreateContactForm
