+++ src/Dashboard/components/NavBar/index.js
@@ -1,0 +1,2 @@
+import NavBar from './NavBar'
+export default NavBar
