+++ src/Dashboard/index.js
@@ -1,0 +1,2 @@
+import DashBoardContainer from './DashboardContainer'
+export default DashBoardContainer
